Adapted colored texts to more Team-contexts
Had to add a prepossessor for p_interaction however.
